1. Commit
Introduction

It is now possible to collect a large amount of data about personal
movement using activity monitoring devices such as a Fitbit, Nike
Fuelband, or Jawbone Up. These type of devices are part of the
“quantified self” movement – a group of enthusiasts who take
measurements about themselves regularly to improve their health, to find
patterns in their behavior, or because they are tech geeks. But these
data remain under-utilized both because the raw data are hard to obtain
and there is a lack of statistical methods and software for processing
and interpreting the data.

This assignment makes use of data from a personal activity monitoring
device. This device collects data at 5 minute intervals through out the
day. The data consists of two months of data from an anonymous
individual collected during the months of October and November, 2012 and
include the number of steps taken in 5 minute intervals each day.

The data for this assignment can be downloaded from the course web site:

Dataset: Activity monitoring data [52K]
The variables included in this dataset are:

steps: Number of steps taking in a 5-minute interval (missing values are
coded as NA)
date: The date on which the measurement was taken in YYYY-MM-DD format
interval: Identifier for the 5-minute interval in which measurement was
taken
The dataset is stored in a comma-separated-value (CSV) file and there
are a total of 17,568 observations in this dataset.

Loading and preprocessing the data

Show any code that is needed to

Load the data (i.e. read.csv())
Process/transform the data (if necessary) into a format suitable for
your analysis
What is mean total number of steps taken per day?

For this part of the assignment, you can ignore the missing values in
the dataset.

Calculate the total number of steps taken per day
If you do not understand the difference between a histogram and a
barplot, research the difference between them. Make a histogram of the
total number of steps taken each day
Calculate and report the mean and median of the total number of steps
taken per day
What is the average daily activity pattern?

Make a time series plot (i.e. type = "l") of the 5-minute interval
(x-axis) and the average number of steps taken, averaged across all days
(y-axis)
Which 5-minute interval, on average across all the days in the dataset,
contains the maximum number of steps?
Imputing missing values

Note that there are a number of days/intervals where there are missing
values (coded as NA). The presence of missing days may introduce bias
into some calculations or summaries of the data.

Calculate and report the total number of missing values in the dataset
(i.e. the total number of rows with NAs)
Devise a strategy for filling in all of the missing values in the
dataset. The strategy does not need to be sophisticated. For example,
you could use the mean/median for that day, or the mean for that
5-minute interval, etc.
Create a new dataset that is equal to the original dataset but with the
missing data filled in.
Make a histogram of the total number of steps taken each day and
Calculate and report the mean and median total number of steps taken per
day. Do these values differ from the estimates from the first part of
the assignment? What is the impact of imputing missing data on the
estimates of the total daily number of steps?
Are there differences in activity patterns between weekdays and
weekends?

For this part the weekdays() function may be of some help here. Use the
dataset with the filled-in missing values for this part.

Create a new factor variable in the dataset with two levels – “weekday”
and “weekend” indicating whether a given date is a weekday or weekend
day.
Make a panel plot containing a time series plot (i.e. type = "l") of the
5-minute interval (x-axis) and the average number of steps taken,
averaged across all weekday days or weekend days (y-axis). See the
README file in the GitHub repository to see an example of what this plot
should look like using simulated data.
Submitting the Assignmentless
To submit the assignment:

Commit your completed PA1_template.Rmd file to the master branch of your
git repository (you should already be on the master branch unless you
created new ones)
Commit your PA1_template.md and PA1_template.html files produced by
processing your R markdown file with knit2html() function in R (from the
knitr package) by running the function from the console.
If your document has figures included (it should) then they should have
been placed in the figure/ directory by default (unless you overrided
the default). Add and commit the figure/ directory to your git
repository so that the figures appear in the markdown file when it
displays on github.
Push your master branch to GitHub.
Submit the URL to your GitHub repository for this assignment on the
course web site.
In addition to submitting the URL for your GitHub repository, you will
need to submit the 40 character SHA-1 hash (as string of numbers from
0-9 and letters from a-f) that identifies the repository commit that
contains the version of the files you want to submit. You can do this in
GitHub by doing the following

Going to your GitHub repository web page for this assignment
Click on the “?? commits” link where ?? is the number of commits you
have in the repository. For example, if you made a total of 10 commits
to this repository, the link should say “10 commits”.
You will see a list of commits that you have made to this repository.
The most recent commit is at the very top. If this represents the
version of the files you want to submit, then just click the “copy to
clipboard” button on the right hand side that should appear when you
hover over the SHA-1 hash. Paste this SHA-1 hash into the course web
site when you submit your assignment. If you don't want to use the most
recent commit, then go down and find the commit you want and copy the
SHA-1 hash.